requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
attrs==23.2.0
black==24.4.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
docker-pycreds==0.4.0
filelock==3.13.1
flake8==7.0.0
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.64.1
hydra-core==1.3.2
idna==3.7
importlib-metadata==7.1.0
Jinja2==3.1.3
lightning==2.2.5
lightning-bolts==0.7.0
lightning-utilities==0.11.2
Markdown==3.6
MarkupSafe==2.1.5
mccabe==0.7.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.7.0.84
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.20.5
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
packaging==24.0
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.2
protobuf==4.25.3
psutil==5.9.8
pycodestyle==2.11.1
pyflakes==3.2.0
pytorch-lightning==1.9.5
PyYAML==6.0.1
requests==2.32.3
sentry-sdk==2.5.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.12
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tomli==2.0.1
torch==2.3.1+cu118
torchaudio==2.3.1+cu118
torchmetrics==1.4.0.post0
torchvision==0.18.1+cu118
tqdm==4.66.4
triton==2.3.1
typing-extensions==4.9.0
urllib3==2.2.1
wandb==0.17.1
werkzeug==3.0.3
yarl==1.9.4
zipp==3.19.2
